rest: use RequestTS for Alarm clearing
On TaskDeprovision requests, validUntil is set by eye to now()+15min.
This is so that the configuration remains valid until the next version
is rolled out.
If a new version of the configuration is expected, usually no
AlarmClearing by Eye is required -- none is done by default.

On TaskDelete requests, validUntil is actually set by eye to now().
Which means that later, when rest.Respond receives the Result there is
no valid configuration any more at then now().

Fix this by using the RequestTS for AlarmClearing, which is the
timestamp when the Request was initially received. This timestamp is
set by BasicAuth when accepting the request and transported through the
entire processing pipeline.
It is thus guaranteed to pre-date the invalidation timestamp used by
eye.