proto.v2: add current timestamp to Snapshot
There are multiple timestamps involved when a deprovision
request is received:

- RequestTS, the timestamp of the request set by the first handler
  processing it (usually rest.BasicAuth)

- TransactionTS, the timestamp used by the application during the
  invalidation of the configuration to ensure seamless histories

- ValidUntilTS, this timestamp is derived from TransactionTS and is
  either equal or TransactionTS+15min to keep external caching hot
  in configuration updates

The return data after the client, ie. the data available to rest on
the outbound path to generate Notification events has ValidUntil
timestamps correctly set.

So to snapshot the configuration, a timestamp is required that is
before both TransactionTS and ValidUntilTS - RequestTS is used.

But the timestamp of the notification event must be current, otherwise
the generated clear could end up older than other events generated
before cache invalidation.

Add field CurrentTS to Snapshot that can be used in templates to
get a current timestamp.